+++ lib/auth/form_submission_status.dart
@@ -1,0 +1,23 @@
+// classes for holding the status of form submissions
+
+abstract class FormSubmissionStatus {
+  const FormSubmissionStatus();
+}
+
+class InitialFormStatus extends FormSubmissionStatus {
+  const InitialFormStatus();
+}
+
+class FormSubmitting extends FormSubmissionStatus {
+
+}
+
+class SubmissionSuccess extends FormSubmissionStatus {
+
+}
+
+class SubmissionFailed extends FormSubmissionStatus {
+  final Exception exception;
+
+  SubmissionFailed(this.exception);
+}
