+++ lib/auth/login/login_event.dart
@@ -1,0 +1,20 @@
+
+abstract class LoginEvent {
+  
+}
+
+class LoginUsernameChanged extends LoginEvent {
+  final String username;
+
+  LoginUsernameChanged({this.username});
+}
+
+class LoginPasswordChanged extends LoginEvent {
+  final String password;
+
+  LoginPasswordChanged({this.password});
+}
+
+class LoginSubmitted extends LoginEvent {
+
+}
